+++ tests/data/numpy/yields/cases.py
@@ -342,7 +342,7 @@ class A:
 
         yield inner9d(arg1)
 
-    def method10(self, n: int) -> Generator[str, None, None]:
+    def method10a(self, n: int) -> Generator[str, None, None]:
         """Description
 
         Parameters
@@ -356,3 +356,18 @@ class A:
             Description
         """
         yield from range(n)
+
+    def method10b(self, n: int) -> Generator[tuple[float, ...], None, None]:
+        """Foo
+
+        Parameters
+        ----------
+        n : int
+            Description.
+
+        Yields
+        ------
+        tuple[float, ...]
+            Description.
+        """
+        yield from ((*self.bar, i) for i in range(self.baz))